Create CopyPasteSave.m
A customer ran into an issue where they would copy and paste a satellite that had astrogator as it's propagator. The new, copied, satellite would run into issues with the scripting tool when the MCS was run however if you saved the scenario before running the MCS the issue would go away. This snippet automates the process of copying, pasting, naming, and saving so that all that needs to be done is change the value of satName to be the name of the satellite you want to be copied.  Once the first two sections have been run once, the final two sections can be run repeatedly so long as they are run together and in sequence.