fixing minor erros in this .py file
I was using STK 11. I was playing around with this script. During this process, I found that function deckAccessReportToDF() would throw an error, which is because the local variable header in this function is empty. I fixed the header issue. Also, the pandas dataframe that the function returns has an extra first line. So I added .iloc[1:] in the end.