Can now easily add bandwidth node constraints
This is useful for limiting bandwidth/flow capacity for a particular node, instead of just a connection/edge.

Also fixed some small off by 1 errors in the DataTransfer notebook. 

Also added a way to filter the network down to only show the routes passing through a specified set of nodes.