Create PrimaryAreaTargetOnly.py
the "GIS Import" command will load all AreaTargets contained in the selected .shp, there is currently no way to specify that only the primary region is desired. This script will load all AreaTargets from the selected shapefile, then loop through all of them to find the one with the largest area, assuming the one with the largest area is the primary region. It will then remove all other areas that were imported from this shp