Fix non-obscured interval generation bug
This change fixes a bug that occurs when mask sample start time is 0 EpSec and access exists. When this happens, the first interval start time gets removed and errors or incorrect time intervals are generated. This fix mitigates this issue by checking if the condition is met and then preserving the first interval if it is.